+++ app/javascript/vue/components/radials/navigation/radial.vue
@@ -419,6 +419,7 @@ function destroyObject() {
         )
       }
     }
+    emit('delete', metadata.value)
   })
 }
 </script>