+++ app/javascript/vue/tasks/biological_associations/filter/app.vue
@@ -43,6 +43,7 @@
           :header-groups="HEADERS"
           :list="list"
           @on-sort="list = $event"
+          @remove="({ index }) => list.splice(index, 1)"
         />
       </template>
     </FilterLayout>