+++ app/javascript/vue/tasks/otu/filter/App.vue
@@ -44,6 +44,7 @@
           v-model="selectedIds"
           radial-object
           @on-sort="list = $event"
+          @remove="({ index }) => list.splice(index, 1)"
         />
       </template>
     </FilterLayout>